Pluginify search aggregation stage (#3817)
