+++ src/azul/service/elasticsearch_service.py
@@ -381,22 +381,6 @@ class AggregationStage(_ElasticsearchStage[MutableJSON, MutableJSON]):
                    field=path,
                    size=config.terms_aggregation_size)
         agg.bucket('untagged', 'missing', field=path)
-        if facet == 'project':
-            sub_path = self.plugin.field_mapping['projectId'] + '.keyword'
-            agg.aggs['myTerms'].bucket(name='myProjectIds',
-                                       agg_type='terms',
-                                       field=sub_path,
-                                       size=config.terms_aggregation_size)
-        elif facet == 'fileFormat':
-            # FIXME: Use of shadow field is brittle
-            #        https://github.com/DataBiosphere/azul/issues/2289
-            def set_summary_agg(field: str, bucket: str) -> None:
-                path = self.plugin.field_mapping[field] + '_'
-                agg.aggs['myTerms'].metric(bucket, 'sum', field=path)
-                agg.aggs['untagged'].metric(bucket, 'sum', field=path)
-
-            set_summary_agg(field='fileSize', bucket='size_by_type')
-            set_summary_agg(field='matrixCellCount', bucket='matrix_cell_count_by_type')
         return agg
 
     def _annotate_aggs_for_translation(self, request: Search):
